removed decorators
+ various cleanups
